Add a barebones homebrew forumla for release
This is an extremely basic attempt at a homebrew formula for release. Please adjust it to suit your needs. Technically it doesn't need to exist in this repo (or anywhere) unless you want it to be a separate tap. When it's in a state you're happy with, we can open a PR against homebrew/core