ui: move SSH Agent thread initialization to ui setup
This ensures that the context is always populated on the state
before we run the agent. If we don't do this, there is a race
between the context being populated and the agent processing a
request that requires access to UI elements.

This should fix an intermittent startup crash due to the
aforementioned race condition.