ui: properly release strings
Before, we didn't have an autorelease pool when reflecting state
updates and we were leaking strings. This resulted in a gradual
memory leak.

After, we attribute strings to an autoreleasepool and we autorelease
them after we're done with them.

According to the `leak` tool, this plugs most of the leaks in this
program.